Moved server code into its own header and source files.
