+++ CMakeLists.txt
@@ -7,6 +7,8 @@ project(
    VERSION 0.0.1
 )
 
+include( ${CMAKE_SOURCE_DIR}/cmake/functions.cmake )
+
 add_compile_definitions(
    NETWORKING_DOODOO
    PRIVATE
@@ -19,68 +21,66 @@ set( EXECUTABLE_OUTPUT_PATH "${CMAKE_BINARY_DIR}/bin" )
 
 # How to copy libraries (DLL's on Windows) over to binary directory
 # https://stackoverflow.com/a/63158941
-set(
-   sources
-   client.cpp
-   # server.cpp
-)
 
-set(
-   headers
+add_subdirectory( demos/client_server )
 
-)
+add_subdirectory( junkwork )
 
-set(
-   target
-   networking_doodoo
-)
-
-add_executable(
-   ${target}
-   ${sources}
-   ${headers}
-)
-
-# add_library(
-#     ${target}
-#     SHARED
-#     ${sources}
-#     ${headers}
+# set(
+#    sources
+#    client.cpp
+#    # server.cpp
 # )
 
-include_directories(
-   ${CMAKE_CURRENT_SOURCE_DIR}
-)
+# set(
+#    headers
 
-target_include_directories(
-   ${target}
-   PUBLIC
-   ${CMAKE_CURRENT_SOURCE_DIR}
-)
+# )
 
-set_target_properties(
-   ${target}
-   PROPERTIES
-   LINKER_LANGUAGE
-   CXX
-)
+# set(
+#    target
+#    networking_doodoo
+# )
 
-set(
-   link_libraries
-)
+# add_executable(
+#    ${target}
+#    ${sources}
+#    ${headers}
+# )
 
-# append wsock2 and ws2_32 to the link libraries variable if on Windows
+# include_directories(
+#    ${CMAKE_CURRENT_SOURCE_DIR}
+# )
 
-target_link_libraries(
-   ${target}
-)
+# target_include_directories(
+#    ${target}
+#    PUBLIC
+#    ${CMAKE_CURRENT_SOURCE_DIR}
+# )
 
-target_compile_options(
-   ${target}
-   PRIVATE
-   -Wall
-   -Wextra
-   -Wpedantic
-   -Wno-unused-function
-   # -Werror
-)
+# set_target_properties(
+#    ${target}
+#    PROPERTIES
+#    LINKER_LANGUAGE
+#    CXX
+# )
+
+# set(
+#    link_libraries
+# )
+
+# # append wsock2 and ws2_32 to the link libraries variable if on Windows
+
+# target_link_libraries(
+#    ${target}
+# )
+
+# target_compile_options(
+#    ${target}
+#    PRIVATE
+#    -Wall
+#    -Wextra
+#    -Wpedantic
+#    -Wno-unused-function
+#    # -Werror
+# )
