+++ demos/client_server/CMakeLists.txt
@@ -1,0 +1,3 @@
+make_executable("client_demo" "client.cpp" "" "junkwork")
+
+# make_executable("server_demo" "server.cpp" "" "junkwork")
