+++ junkwork/include/junkwork/basic_udp.hpp
@@ -1,0 +1,51 @@
+#define PLATFORM_WINDOWS 1
+#define PLATFORM_MAC     2
+#define PLATFORM_UNIX    3
+
+#if defined(_WIN32)
+#define PLATFORM PLATFORM_WINDOWS
+#elif defined(__APPLE__)
+#define PLATFORM PLATFORM_MAC
+#else
+#define PLATFORM PLATFORM_UNIX
+#endif
+
+#if PLATFORM == PLATFORM_WINDOWS
+
+#include <winsock2.h>
+
+#elif PLATFORM == PLATFORM_MAC || \
+      PLATFORM == PLATFORM_UNIX
+
+#include <sys/socket.h>
+#include <netinet/in.h>
+#include <fcntl.h>
+#include <unistd.h>
+
+#endif
+
+bool initialize_sockets(void);
+
+void shutdown_sockets(void);
+
+int create_socket(void);
+
+bool bind_socket(const int handle, const unsigned int port);
+
+void close_socket(
+   const int handle
+);
+
+bool send_packet(
+   const int handle,
+   const sockaddr_in to_address,
+   const char * data,
+   const unsigned int packet_size
+);
+
+bool receive_packet(
+   const int socket,
+   const unsigned int max_packet_size,
+   sockaddr_in * from_address,
+   char * packet_data
+);
