+++ junkwork/src/basic_udp.cpp
@@ -1,28 +1,4 @@
-#define PLATFORM_WINDOWS 1
-#define PLATFORM_MAC     2
-#define PLATFORM_UNIX    3
-
-#if defined(_WIN32)
-#define PLATFORM PLATFORM_WINDOWS
-#elif defined(__APPLE__)
-#define PLATFORM PLATFORM_MAC
-#else
-#define PLATFORM PLATFORM_UNIX
-#endif
-
-#if PLATFORM == PLATFORM_WINDOWS
-
-#include <winsock2.h>
-
-#elif PLATFORM == PLATFORM_MAC || \
-      PLATFORM == PLATFORM_UNIX
-
-#include <sys/socket.h>
-#include <netinet/in.h>
-#include <fcntl.h>
-#include <unistd.h>
-
-#endif
+#include "junkwork/basic_udp.hpp"
 
 #include <iostream>
 
@@ -162,68 +138,4 @@ bool receive_packet(
    );
 
    return bytes > 0;
-}
-
-int main(int argc, char ** argv)
-{
-   if (argc < 2)
-   {
-      return 0;
-   }
-
-   // create the socket
-   int handle = create_socket();
-
-   // // bind the socket
-   // unsigned int port = 8123;
-
-   unsigned int bind_port = atoi(argv[1]);
-
-   std::cout << "port: " << bind_port << "\n";
-
-   if (!bind_socket(handle, bind_port))
-   {
-      return 1;
-   }
-
-   unsigned int a = 127;
-   unsigned int b = 0;
-   unsigned int c = 0;
-   unsigned int d = 1;
-
-   unsigned int ip_add = (a << 24) | (b << 16) | ( c << 8) | d;
-
-   sockaddr_in send_address;
-   send_address.sin_family = AF_INET;
-   send_address.sin_addr.s_addr = htonl(ip_add);
-   send_address.sin_port = htons(8123);
-
-   send_packet(handle, send_address, "stuff", sizeof("stuff"));
-
-   if (bind_port != 8123)
-   {
-      return 0;
-   }
-
-   sockaddr_in from_address;
-   char packet_data[6];
-
-   while(true)
-   {
-      // if (receive_packet(handle, 6, &from_address, packet_data))
-      if (receive_packet(handle, 6, &from_address, packet_data))
-      {
-         std::cout << "packet: " << packet_data << "\n";
-      }
-   }
-
-   // std::cout << "got packet\n";
-
-   // for (int i = 0; i < 5; ++i)
-   // {
-   //    std::cout << static_cast<int>(packet_data[i]) << " ";
-   // }
-   // std::cout << "\n";
-
-   return 0;
 }
